pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = ""
authors = [
    {name = "sjoerdoffringa",email = "sjoerd.offringa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "numpy (>=2.2.4,<3.0.0)",
    "pypdf (>=5.4.0,<6.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "langchain (>=0.3.22,<0.4.0)",
    "sentence-transformers (>=4.0.1,<5.0.0)",
    "rerankers[flashrank] (>=0.9.1.post1,<0.10.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "blobfile (>=3.0.0,<4.0.0)",
    "datasets (>=3.5.0,<4.0.0)",
    "uv (>=0.6.14,<0.7.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
matplotlib = "^3.10.1"
mlflow = "^2.21.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
